requirements.txt
ansible==8.7.0
PyYAML>=6.0
